Clean up Module::Install usage.
Remove Module::Install from the repo, since it will package itself, as
needed when doing a "make dist".

Update the .gitignore file to make sure we don't accidentally add
and dist, or Module::Install files.

Create MANIFEST.SKIP to keep the .git directory from getting included in
the MANIFEST, along with a few other files that really shouldn't be put in
the MANIFEST.